Switched new naming `LLVMcore` to existing `LLVM`
